completed dialogfragments for getting user input for recipes and ingredients, and the number of each they want, for the meal plan

+++ app/src/main/java/com/androidimpact/app/meal_plan/MealAdapter.java
@@ -26,7 +26,7 @@ public class MealAdapter extends RecyclerView.Adapter<MealAdapter.MealViewHolder
     private MealPlan mealPlan;
     ArrayList<Recipe> recipeArrayList;
     ArrayList<StoreIngredient> ingredientArrayList;
-    ArrayList<Float> recipeServingsArrayList, ingredientServingsArrayList;
+    ArrayList<Double> recipeServingsArrayList, ingredientServingsArrayList;
 
 
     String key;
